gsocketlistener: Fix memory corruption
The implementation of g_socket_listener_accept_socket_async() assumed
that the task would get destroyed before returning from accept_ready(),
and hence it could just tie the life-time of the sources to the life-
time of the task, and drop the task reference before returning from
accept_ready(). However, if an idle source had to be scheduled in order
to deliver the task's result, this meant that the task, and consequently
also its sources, would stay alive a bit longer. By doing so, additional
callbacks to accept_ready() could happen in the meantime, i.e. more than
one socket might be ready for dispatching, and it would start dropping
references it didn't own.