gwin32: Add GWin32Dispatcher for message pump integration
Integrating an external main loop with the Windows message pump is
anything but trivial. This is primarily because DispatchMessage()
may block, leaving GMainContext to starve indefinitely. The most
common example is in a UI application while resizing a toplevel
window.  Windows will in this case, as part of its message handling
when the resize is started, start a nested message pump which processes
messages until the user chooses to let go in order to accept the new
window size.