gsocketlistener: Fix IPv4 listen() error-handling
In add_any_inet_port() for the case where we successfully listen()ed
on IPv6, but failed to listen on IPv4, we would erroneously unref the
IPv6 socket which we just gave away ownership of by adding it to
priv->sockets.

However, given that the existing logic is best-effort we shouldn't
report it as an error if IPv4 listen() fails in the case where we
actually managed to listen() on IPv6.

This issue was observed in an Android 4.4 ARM emulator.

https://bugzilla.gnome.org/show_bug.cgi?id=780359