meson: Do not treat unused-result warning as error
To fix compilation errors when building against libc headers where
posix_memalign() is declared with `warn_unused_result`.

This might already be fixed upstream, but for Frida's use-case we don't
handle out-of-memory situations elsewhere so it's not worthwhile to
maintain a patch that only does it for this particular case.

Fixes #10